set-allert

+++ app/Http/Livewire/Divisi/DivisiAllComponent.php
@@ -7,12 +7,29 @@ use App\Models\Divisi;
 
 class DivisiAllComponent extends Component
 {
+    public $divisid;
 
-     public function delete($id)
+    protected $listeners = ['deleteConfirmed' => 'deleteDivisi'];
+
+    public function mount()
+    {
+        $this->divisid = Divisi::all();
+    }
+
+    public function confirmDelete($id)
+    {
+        $this->dispatchBrowserEvent('show-delete-confirmation', ['id' => $id]);
+    }
+
+
+     public function deleteDivisi($id)
     {
         $divisi = Divisi::find($id);
-        $divisi->delete();
-        session()->flash('notif', 'berhasil dihapus');
+        if ($divisi){
+            $divisi->delete();
+            session()->flash('notif', 'berhasil dihapus');
+            $this->divisid = Divisi::all();
+        }
     }
 
     public function render()
